Separate simulation from general Pool testing
